Smoketest hdf5_file_info()
